% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_funcs.R
\name{copy_files_jitter}
\alias{copy_files_jitter}
\title{Copy files to initialize jitter}
\usage{
copy_files_jitter(dir, jitinds, jitamount)
}
\arguments{
\item{dir}{Directory containing the jitter runs}

\item{jitinds}{Number of jitter runs}

\item{jitamount}{Percentage to jitter runs; default is 0.05 or 5 percent.}
}
\description{
Copy jitter files, directory should contain a folder "0" that has the base model
}
\author{
Peter Kuriyama
}
