% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{pull_pars_lls}
\alias{pull_pars_lls}
\title{Pull the parameters and total likelihood}
\usage{
pull_pars_lls(res)
}
\arguments{
\item{res}{List of results for evaluation}
}
\description{
Pull the parameters and total likelihood
}
