% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{pull_timeseries}
\alias{pull_timeseries}
\title{Pull the time series values}
\usage{
pull_timeseries(res)
}
\arguments{
\item{res}{List of results for evaluation}
}
\description{
Pull the time series values
}
