% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_funcs.R
\name{rr_jitter_parallel}
\alias{rr_jitter_parallel}
\title{Run jitters in parallel}
\usage{
rr_jitter_parallel(
  thedir = "00_caal/4a_cohort_seprofile/7_jitters/",
  thecores,
  thejitinds = 1:8,
  thejitamount,
  SScall
)
}
\arguments{
\item{thedir}{Directory containing the jitter runs}

\item{thecores}{Number of cores to run in parallel}

\item{thejitamount}{Percentage to jitter runs; default is 0.05 or 5 percent.}

\item{SScall}{Directory of Stock Synthesis Call, make sure it has -nohess}

\item{jitinds}{Number of jitter runs}
}
\description{
Run SS jitter runs with specified number of cores
}
\author{
Peter Kuriyama
}
