% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ss_parallel.R
\name{ssoutput_parallel}
\alias{ssoutput_parallel}
\title{Read SS parallel}
\usage{
ssoutput_parallel(ncores = 8, folders)
}
\arguments{
\item{ncores}{Number of cores; default to 8}

\item{inds}{Indices of parallel runs; applies only to jitters}

\item{syscall}{Point to desired version of SS}

\item{dir}{Directory of jitter runs}

\item{readres}{Read in jitter results if TRUE, don't if FALSE}
}
\description{
across a number of cores
}
\author{
Peter Kuriyama
}
